// appfront2/src/interfaces/remi-config-data.ts
export interface RemiConfigData {
  initialWaitTime?: number;
  automaticReactivate: boolean;
  initialMessage?: string;
  automaticWaitTime?: number;
  automaticMessage?: string;
  finalizationWaitTime?: number;
  finalizationMessage?: string;
  name?: string;
  teamIds?: string[];
  id?: string;
  active?: boolean;
  objectiveId?: number;
  outcomeId?: number;
}